Refactoring of evaluators -> moved to tests directory
