+++ xopt/tests/test_cnsga.py
@@ -12,7 +12,7 @@ algorithm:
   
 simulation: 
   name: test_TNK
-  evaluate: xopt.evaluators.test_TNK.evaluate_TNK  
+  evaluate: xopt.tests.evaluators.TNK.evaluate_TNK  
   
 vocs:
   name: TNK_test
